requirements.txt
keras==2.7.0
matplotlib==3.5.1
numpy==1.21.4
Pillow==8.4.0
tensorflow==2.7.0
